# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THE HOMIES ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # USERS
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List partners
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{id}:
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    delete:
      operationId: deletePartner
      tags: [partners]
      summary: Delete partner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /partners/send-product:
    post:
      operationId: sendProductToPartner
      tags: [partners]
      summary: Send product to partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerSendInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerMovement"
  /partners/return-product:
    post:
      operationId: returnProductFromPartner
      tags: [partners]
      summary: Return product from partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerReturnInput"
      responses:
        "200":
          description: Returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerMovement"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CLIENTS
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients (admin only)
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client with purchase history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # MOVEMENTS (Journal)
  /movements:
    get:
      operationId: listMovements
      tags: [movements]
      summary: List all movements
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movement"

  # STATISTICS (admin only)
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, username, fullName, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, secretary]
        createdAt:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, productId, product, brand, status, entryDate, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: string
          description: "Auto-generated like TH001"
        imei:
          type: ["string", "null"]
        product:
          type: string
        brand:
          type: string
        capacity:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        supplier:
          type: ["string", "null"]
        purchasePrice:
          type: ["number", "null"]
          description: "Hidden from secretary"
        sellingPrice:
          type: ["number", "null"]
        profit:
          type: ["number", "null"]
          description: "Calculated: sellingPrice - purchasePrice"
        status:
          type: string
          enum: [en_stock, chez_partenaire, vendu]
        entryDate:
          type: string
          format: date
        saleDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
...
    Sale:
      type: object
      required: [id, productId, saleType, paymentMode, amount, saleDate, saleTime, sellerId, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        saleType:
          type: string
          enum: [normal, troc]
        paymentMode:
          type: string
          enum: [OM, MOMO, Cash]
        amount:
          type: number
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        sellerId:
          type: integer
        seller:
          $ref: "#/components/schemas/User"
        saleDate:
          type: string
          format: date
        saleTime:
          type: string
        cancelled:
          type: boolean
        cancellationReason:
          type: ["string", "null"]
        trocProductId:
          type: ["integer", "null"]
          description: "ID of product received in troc"
        createdAt:
          type: string
          format: date-time
...
    Partner:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PartnerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
    PartnerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
    PartnerMovement:
      type: object
      required: [id, partnerId, productId, movementType, movementDate, createdAt]
      properties:
        id:
          type: integer
        partnerId:
          type: integer
        partner:
          $ref: "#/components/schemas/Partner"
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        movementType:
          type: string
          enum: [sortie, retour]
        movementDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PartnerSendInput:
      type: object
      required: [partnerId, productId, movementDate]
      properties:
        partnerId:
          type: integer
        productId:
          type: integer
        movementDate:
          type: string
          format: date
        notes:
          type: string
    PartnerReturnInput:
      type: object
      required: [productId, movementDate]
      properties:
        productId:
          type: integer
        movementDate:
          type: string
          format: date
        notes:
          type: string

    # EXPENSE
    Expense:
      type: object
      required: [id, label, amount, expenseDate, expenseTime, userId, createdAt]
      properties:
        id:
          type: integer
        label:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
          format: date
        expenseTime:
          type: string
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [label, amount, expenseDate]
      properties:
        label:
          type: string
        amount:
          type: number
        expenseDate:
          type: string
          format: date
    ExpenseUpdate:
      type: object
      properties:
        label:
          type: string
        amount:
          type: number
        expenseDate:
          type: string

    # CLIENT
    Client:
      type: object
      required: [id, fullName, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        purchaseCount:
          type: integer
        totalPurchases:
          type: number
        lastPurchaseDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    ClientDetail:
      type: object
      required: [id, fullName, createdAt, purchases]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        purchaseCount:
          type: integer
        totalPurchases:
          type: number
        lastPurchaseDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        purchases:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
    ClientInput:
      type: object
      required: [fullName]
      properties:
        fullName:
          type: string
        phone:
          type: string
    ClientUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string

    # MOVEMENT (Journal)
    Movement:
      type: object
      required: [id, movementType, movementDate, movementTime, userId, description, createdAt]
      properties:
        id:
          type: integer
        movementType:
          type: string
          enum: [achat, vente, entree_troc, depense, sortie_partenaire, retour_partenaire, modification_produit, suppression_produit, annulation]
        movementDate:
          type: string
          format: date
        movementTime:
          type: string
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        productId:
          type: ["integer", "null"]
        productRef:
          type: ["string", "null"]
        imei:
          type: ["string", "null"]
        description:
          type: string
        createdAt:
          type: string
          format: date-time

    # STATS
...