# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THE HOMIES ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  label?: string;
  amount?: number;
  expenseDate?: string;
}